+++ modules/dreamview/backend/simulation_world/simulation_world_service.cc
@@ -944,9 +944,10 @@ void SimulationWorldService::UpdatePlanningData(const PlanningData &data) {
   }
 
   // Update pull over status
+  planning_data->clear_pull_over_status();
   if (data.has_pull_over_status()) {
-    planning_data->clear_pull_over_status();
-    planning_data->mutable_pull_over_status()->CopyFrom(data.pull_over_status());
+    planning_data->mutable_pull_over_status()->CopyFrom(
+      data.pull_over_status());
   }
 }
