+++ modules/dreamview/frontend/src/renderer/trajectory.js
@@ -1,7 +1,6 @@
 import _ from 'lodash';
 import STORE from "store";
-import * as THREE from "three";
-import { drawThickBandFromPoints, drawSegmentsFromPoints, drawCircle, drawArrow } from "utils/draw";
+import { drawThickBandFromPoints } from "utils/draw";
 
 function normalizePlanningTrajectory(trajectory, coordinates) {
     if (!trajectory) {
@@ -37,42 +36,9 @@ function normalizePlanningTrajectory(trajectory, coordinates) {
     return result;
 }
 
-function getPullOverStatus({ lengthFront, lengthBack, widthLeft, widthRight }) {
-    const pullOverStatus = new THREE.Group();
-    const color = 0x006AFF;
-    const polygon = drawSegmentsFromPoints(
-        [
-            new THREE.Vector3(lengthFront, -widthLeft, 0),
-            new THREE.Vector3(lengthFront, widthRight, 0),
-            new THREE.Vector3(-lengthBack, widthRight, 0),
-            new THREE.Vector3(-lengthBack, -widthLeft, 0),
-            new THREE.Vector3(lengthFront, -widthLeft, 0),
-        ],
-        color,
-        2,
-        5,
-    );
-    pullOverStatus.add(polygon);
-
-    const material = new THREE.MeshBasicMaterial({
-        color,
-        transparent: false,
-        opacity: 0.5
-    });
-    const circle = drawCircle(0.2, material);
-    pullOverStatus.add(circle);
-
-    const heading = drawArrow(1.5, 2, 0.5, 0.5, color);
-    heading.rotation.set(0, 0, -Math.PI / 2);
-    pullOverStatus.add(heading);
-
-    return pullOverStatus;
-}
-
 export default class PlanningTrajectory {
     constructor() {
         this.paths = {};
-        this.pullOverStatus = null;
     }
 
     update(world, planningData, coordinates, scene) {
@@ -131,39 +97,5 @@ export default class PlanningTrajectory {
             }
             propertyIndex += 1;
         });
-
-        // Draw pull over status
-        if (planningData) {
-            const { pullOverStatus } = planningData;
-            if (!STORE.options.customizedToggles.get('pullOverStatus')) {
-                if (this.pullOverStatus) {
-                    this.pullOverStatus.traverse((child) => {
-                        child.visible = false;
-                    });
-                }
-            } else {
-                if (this.pullOverStatus) {
-                    this.pullOverStatus.traverse((child) => {
-                        if (child.geometry !== undefined) {
-                            child.geometry.dispose();
-                            child.material.dispose();
-                        }
-                        scene.remove(this.pullOverStatus);
-                    });
-                }
-                if (pullOverStatus) {
-                    this.pullOverStatus = getPullOverStatus(pullOverStatus);
-
-                    const position = coordinates.applyOffset({
-                        x: pullOverStatus.position.x,
-                        y: pullOverStatus.position.y,
-                        z: 0.3,
-                    });
-                    this.pullOverStatus.position.set(position.x, position.y, position.z);
-                    this.pullOverStatus.rotation.set(0, 0, pullOverStatus.theta);
-                    scene.add(this.pullOverStatus);
-                }
-            }
-        }
     }
 }
