+++ modules/dreamview/frontend/src/store/index.js
@@ -170,35 +170,30 @@ class DreamviewStore {
     }
 
     updateCustomizedToggles(world) {
-        const previousToggles = new Map(this.options.customizedToggles);
+        const newToggles = {};
         if (world.planningData) {
             // Add customized toggles for planning paths
             if (world.planningData.path) {
                 world.planningData.path.forEach((path) => {
                     const pathName = path.name;
-                    if (!previousToggles.has(pathName)) {
-                        this.options.addCustomizedToggle(pathName);
+                    if (this.options.customizedToggles.has(pathName)) {
+                        newToggles[pathName] = this.options.customizedToggles.get(pathName);
                     } else {
-                        previousToggles.delete(pathName);
+                        newToggles[pathName] = true;
                     }
                 });
             }
             // Add pull over status toggle
             if (world.planningData.pullOverStatus) {
                 const keyword = 'pullOverStatus';
-                if (!previousToggles.has(keyword)) {
-                    this.options.addCustomizedToggle(keyword);
+                if (this.options.customizedToggles.has(keyword)) {
+                    newToggles[keyword] = this.options.customizedToggles.get(keyword);
                 } else {
-                    previousToggles.delete(keyword);
+                    newToggles[keyword] = true;
                 }
             }
-            for (const toggle of previousToggles.keys()) {
-                this.options.deleteCustomizedToggle(toggle);
-            }
-        } else if (this.options.customizedToggles.size > 0) {
-            // Clean the planning paths
-            this.options.customizedToggles.clear();
         }
+        this.options.setCustomizedToggles(newToggles);
     }
 
     handleDrivingModeChange(wasAutoMode, isAutoMode) {
